Fixed Previous errors
Currently, the logic needs to be set for WILD card. It should prompt the user to change the color (which needs to be added). Additionally, WILD cards are currently displayed as:
WILD WILD
This should be displayed as:
WILD
So this needs fixing as well.

There might be more logical errors which need debugging.